requirements.txt
Flask==2.1.1
SQLAlchemy==1.4.27

psycopg2-binary==2.9.3

openai==0.27.0
alembic==1.7.7
pytest==6.2.5
python-dotenv==0.19.2
Werkzeug==2.0.1
Flask-SQLAlchemy==2.5.1



